SDK-192: Prepare 3.8.1

+++ CleverTapSDK/ABTesting/uieditor/transformers/CTNSDictionaryFromUIImageValueTransformer.m
@@ -24,7 +24,7 @@
     NSDictionary *imageDictionary = imagesArray[0];
     
     NSNumber *scale = imageDictionary[@"scale"];
-    if (scale <= 0) {
+    if (scale.floatValue <= 0) {
         scale = @(1);
     }
     if (imageDictionary[@"url"]) {
@@ -86,24 +86,24 @@
         }
         
         transformedValue = @{
-                             @"imageOrientation": @(image.imageOrientation),
-                             @"size": [sizeTransformer transformedValue:size],
-                             @"renderingMode": @(image.renderingMode),
-                             @"resizingMode": @(image.resizingMode),
-                             @"duration": @(image.duration),
-                             @"capInsets": [insetsTransformer transformedValue:capInsets],
-                             @"alignmentRectInsets": [insetsTransformer transformedValue:alignmentRectInsets],
-                             @"images": [imageDictionaries copy],
-                             };
+            @"imageOrientation": @(image.imageOrientation),
+            @"size": [sizeTransformer transformedValue:size],
+            @"renderingMode": @(image.renderingMode),
+            @"resizingMode": @(image.resizingMode),
+            @"duration": @(image.duration),
+            @"capInsets": [insetsTransformer transformedValue:capInsets],
+            @"alignmentRectInsets": [insetsTransformer transformedValue:alignmentRectInsets],
+            @"images": [imageDictionaries copy],
+        };
     }
     
     return transformedValue;
 }
 
 - (id)reverseTransformedValue:(id)value {
-     if (![value isKindOfClass:[NSDictionary class]]) {
-          return nil;
-     }
+    if (![value isKindOfClass:[NSDictionary class]]) {
+        return nil;
+    }
     NSDictionary *dictionary = value;
     return [[self class] imageFromDictionary:dictionary];
 }
